+++ torch/distributed/fsdp/_common_utils.py
@@ -32,6 +32,7 @@ from torch.distributed.algorithms._checkpoint.checkpoint_wrapper import (
     _CHECKPOINT_PREFIX,
 )
 from torch.distributed.utils import _apply_to_tensors
+from torch.utils._mode_utils import no_dispatch
 
 from .api import (
     FullOptimStateDictConfig,
@@ -485,3 +486,15 @@ def _override_module_mixed_precision(
             mod.register_forward_pre_hook(forward_pre_hook, prepend=False)
             mod.register_forward_hook(forward_post_hook, prepend=False)
     return overridden_module_classes
+
+
+def _no_dispatch_record_stream(tensor: torch.Tensor, stream: torch.Stream) -> None:
+    # FIXME record_stream doesn't work with non-cuda tensors
+    if tensor.device.type not in ["cuda", torch._C._get_privateuse1_backend_name()]:
+        return
+    with no_dispatch():
+        tensor.record_stream(stream)
+
+
+def _same_storage_as_data_ptr(x: torch.Tensor, data_ptr: int) -> bool:
+    return x._typed_storage()._data_ptr() == data_ptr
